Fix clippy errors

+++ src/main.rs
@@ -36,7 +36,7 @@ fn main() -> ! {
     let tc0 = peripherals.TC0;
     tc0.tccr0a.write(|w| w.wgm0().ctc());
     tc0.tccr0b.write(|w| w.cs0().prescale_8());
-    tc0.ocr0a.write(|w| unsafe { w.bits(124 as u8) });
+    tc0.ocr0a.write(|w| unsafe { w.bits(124_u8) });
     tc0.timsk.write(|w| w.ocie0a().bit(true));
 
     // Enable pin change interrupt for PB0 and PB1 to detect switch changes
